requirements.txt
fastapi
strawberry-graphql
redis
networkx
httpx
uvicorn
matplotlib
apscheduler